fixed bug related to int size

+++ wrapTest/factorial.c
@@ -1,8 +1,8 @@
 #include "factorial.h"
 
-int factorial(int X) {
+long int factorial(int X) {
 	int x;
-	int result = 1, intermediate;
+	long int result = 1, intermediate;
 
 	for (x = 1; x <= X; x++) {
 		intermediate = result * x;
